Backend/estoque/urls.py
from django.urls import path, include
from rest_framework.routers import DefaultRouter
from . import views

router = DefaultRouter()
router.register(r'frutas', views.FrutaViewSet)
router.register(r'categorias', views.CategoriaViewSet)

urlpatterns = [
    path('', include(router.urls)),
    path('dashboard/', views.dashboard, name='dashboard'),
]
